LPD-69528 Migrate playwright tests
