+++ modules/test/playwright/tests/site-dsr-site-initializer/main/dependencies/liferay.png
